// client/src/app/shared/models/interfaces/triangle.interface.ts
export interface TriangleInterface {
  side_a: number;
  side_b: number;
  side_c: number;
  angle_a: number;
  angle_b: number;
  angle_c: number;
  bisector_a: number;
  bisector_b: number;
  bisector_c: number;
  altitude_a: number;
  altitude_b: number;
  altitude_c: number;
  median_a: number;
  median_b: number;
  median_c: number;
  radius_R: number;
  radius_r: number;
  radius_ra: number;
  radius_rb: number;
  radius_rc: number;
}